test(css): flatten css-module-lexer tests
