+++ crates/css-module-lexer/src/dependency_types.rs
@@ -655,5 +655,5 @@ impl Display for Warning<'_> {
 }
 
 #[cfg(test)]
-#[path = "../tests/unit/dependency_types_tests.rs"]
+#[path = "../tests/dependency_types_tests.rs"]
 mod tests;
