+++ crates/css-module-lexer/tests/dependency_types_tests.rs
