--- crates/css-module-lexer/tests/postcss_modules/mod.rs
@@ -1,3 +1,0 @@
-mod extract_imports_test;
-mod local_by_default_test;
-mod scope_test;
