+++ crates/css-module-lexer/tests/scope.rs
@@ -126,7 +126,7 @@ impl Scope {
   }
 }
 
-fn test(input: &str, expected: &str) {
+pub(crate) fn test(input: &str, expected: &str) {
   let input = strip_whitespace_markers(input);
   let expected = strip_whitespace_markers(expected);
   let (actual, warnings) = Scope.transform(&input);
@@ -134,7 +134,7 @@ fn test(input: &str, expected: &str) {
   similar_asserts::assert_eq!(expected, actual);
 }
 
-fn test_with_warning(input: &str, expected: &str, warning: &str) {
+pub(crate) fn test_with_warning(input: &str, expected: &str, warning: &str) {
   let input = strip_whitespace_markers(input);
   let expected = strip_whitespace_markers(expected);
   let (actual, warnings) = Scope.transform(&input);
@@ -145,7 +145,6 @@ fn test_with_warning(input: &str, expected: &str, warning: &str) {
   );
   similar_asserts::assert_eq!(expected, actual);
 }
-
 #[test]
 fn at_rule() {
   test(
@@ -870,837 +869,6 @@ fn error_when_attribute_is_type() {
 
             :export {
                 exportName1: _input__exportName1;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn escape_sequence() {
-  test(
-    indoc! {r#"
-            :local(.smile) {
-                color: red;
-            }
-
-            :local(.smile) :local(.smile) {
-                color: red;
-            }
-
-            :local(.smile) :local(.smile) :local(.smile) {
-                color: red;
-            }
-
-            :local(.smile_with_A) {
-                color: red;
-            }
-
-            .\1F600  :local(.smile) {
-                color: red;
-            }
-
-            :local(.smile) .\1F600 {
-                color: red;
-            }
-
-            .\1F600  :local(.smile) .\1F600 {
-                color: red;
-            }
-
-            .\1F600  :local(.smile_with_A) .\1F600 {
-                color: red;
-            }
-
-            #\1F600  :local(#smile) #\1F600 {
-                color: red;
-            }
-
-            #\1F600  :local(#smile_with_A) #\1F600 {
-                color: red;
-            }
-
-            .a :local(.smile) b {
-                color: red;
-            }
-
-            :local(.smile) > :local(.smile) > :local(.smile) {
-                color: red;
-            }
-
-            .\1F600 :local(.smile) {
-                color: red;
-            }
-
-            .\1F600:local(.smile) {
-                color: red;
-            }
-
-            .\1F600  :local(.smile) {
-                color: red;
-            }
-
-            :local(.smile) .a {
-                color: red;
-            }
-
-            :local(.smile).a {
-                color: red;
-            }
-
-            .a :local(.smile) {
-                color: red;
-            }
-
-            .a:local(.smile) {
-                color: red;
-            }
-        "#},
-    indoc! {r#"
-            ._input__smile {
-                color: red;
-            }
-
-            ._input__smile ._input__smile {
-                color: red;
-            }
-
-            ._input__smile ._input__smile ._input__smile {
-                color: red;
-            }
-
-            ._input__smile_with_A {
-                color: red;
-            }
-
-            .\1F600  ._input__smile {
-                color: red;
-            }
-
-            ._input__smile .\1F600 {
-                color: red;
-            }
-
-            .\1F600  ._input__smile .\1F600 {
-                color: red;
-            }
-
-            .\1F600  ._input__smile_with_A .\1F600 {
-                color: red;
-            }
-
-            #\1F600  #_input__smile #\1F600 {
-                color: red;
-            }
-
-            #\1F600  #_input__smile_with_A #\1F600 {
-                color: red;
-            }
-
-            .a ._input__smile b {
-                color: red;
-            }
-
-            ._input__smile > ._input__smile > ._input__smile {
-                color: red;
-            }
-
-            .\1F600 ._input__smile {
-                color: red;
-            }
-
-            .\1F600._input__smile {
-                color: red;
-            }
-
-            .\1F600  ._input__smile {
-                color: red;
-            }
-
-            ._input__smile .a {
-                color: red;
-            }
-
-            ._input__smile.a {
-                color: red;
-            }
-
-            .a ._input__smile {
-                color: red;
-            }
-
-            .a._input__smile {
-                color: red;
-            }
-
-            :export {
-                smile_with_A: _input__smile_with_A;
-                smile: _input__smile;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_child_class() {
-  test(
-    indoc! {r#"
-            :local(.simple) {
-                color: red;
-            }
-
-            :local(.simple) h1 {
-                color: blue;
-            }
-        "#},
-    indoc! {r#"
-            ._input__simple {
-                color: red;
-            }
-
-            ._input__simple h1 {
-                color: blue;
-            }
-
-            :export {
-                simple: _input__simple;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_class_attribute() {
-  // should be ._input__exportName2[class=_input__exportName1]
-  // but in css-loader after transformed by local_by_default
-  // :local(.exportName2[class="exportName1"]) will become :local(.exportName2)[class="exportName1"]
-  // so the result of css-loader is same with us
-  test(
-    indoc! {r#"
-            :local(.exportName1) {
-                color: red;
-            }
-
-            :local(.exportName2) {
-                color: green;
-            }
-
-            :local(.exportName2[class="exportName1"]) {
-                color: blue;
-            }
-        "#},
-    indoc! {r#"
-            ._input__exportName1 {
-                color: red;
-            }
-
-            ._input__exportName2 {
-                color: green;
-            }
-
-            ._input__exportName2[class="exportName1"] {
-                color: blue;
-            }
-
-            :export {
-                exportName1: _input__exportName1;
-                exportName2: _input__exportName2;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_class_path() {
-  test(
-    indoc! {r#"
-            :local(.exportName) {
-                color: green;
-            }
-        "#},
-    indoc! {r#"
-            ._input__exportName {
-                color: green;
-            }
-
-            :export {
-                exportName: _input__exportName;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_difficult() {
-  // should rename :local() in animation, but using :local() in property is not valid css
-  test(
-    indoc! {r#"
-            @keyframes :local(fade-in) {
-                from {
-                    opacity: 0;
-                }
-            }
-            __CSS_MODULE_LEXER_WHITESPACE__
-            @-webkit-keyframes :local(fade-out) {
-                to {
-                    opacity: 0;
-                }
-            }
-            __CSS_MODULE_LEXER_WHITESPACE__
-            :local(.fadeIn) {
-                animation: :local(fade-in) 5s,
-                :local(fade-out) 1s :local(wrong);
-                content: :local(fade-in), wrong, "difficult, :local(wrong)" :local(wrong);
-            }
-        "#},
-    indoc! {r#"
-            @keyframes _input__fade-in {
-                from {
-                    opacity: 0;
-                }
-            }
-
-            @-webkit-keyframes _input__fade-out {
-                to {
-                    opacity: 0;
-                }
-            }
-
-            ._input__fadeIn {
-                animation: :local(fade-in) 5s,
-                :local(fade-out) 1s :local(wrong);
-                content: :local(fade-in), wrong, "difficult, :local(wrong)" :local(wrong);
-            }
-
-            :export {
-                fade-in: _input__fade-in;
-                fade-out: _input__fade-out;
-                fadeIn: _input__fadeIn;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_global_class() {
-  // we don't support exportGlobals
-  test(
-    indoc! {r#"
-            .exportName {
-                color: green;
-            }
-
-            .exportName:hover {
-                color: red;
-            }
-
-            @media screen {
-                body {
-                    background: red;
-                }
-            }
-
-            :local(.testLocal) {
-                color: blue;
-            }
-        "#},
-    indoc! {r#"
-            .exportName {
-                color: green;
-            }
-
-            .exportName:hover {
-                color: red;
-            }
-
-            @media screen {
-                body {
-                    background: red;
-                }
-            }
-
-            ._input__testLocal {
-                color: blue;
-            }
-
-            :export {
-                testLocal: _input__testLocal;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_global_id() {
-  test(
-    indoc! {r#"
-            #exportName {
-                color: green;
-            }
-
-            #exportName:hover {
-                color: red;
-            }
-
-            @media screen {
-                #exportName-2 {
-                    background: red;
-                }
-            }
-
-            :local(#exportName-3) {
-                color: green;
-            }
-        "#},
-    indoc! {r#"
-            #exportName {
-                color: green;
-            }
-
-            #exportName:hover {
-                color: red;
-            }
-
-            @media screen {
-                #exportName-2 {
-                    background: red;
-                }
-            }
-
-            #_input__exportName-3 {
-                color: green;
-            }
-
-            :export {
-                exportName-3: _input__exportName-3;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_keyframes() {
-  test(
-    indoc! {r#"
-            @keyframes :local(fade-in) {
-                from {
-                    opacity: 0;
-                }
-                100% {
-                    opacity: 1;
-                }
-            }
-
-            @keyframes fade {
-                from {
-                    opacity: 0.5;
-                }
-            }
-
-            :local(.fadeIn) {
-                animation-name: :local(fade-in);
-            }
-
-            :local(.fadeIn) {
-                animation: 2s :local(fade-in);
-            }
-
-            :local(.fadeIn) {
-                animation: :local(fade-in) 2s;
-            }
-        "#},
-    indoc! {r#"
-            @keyframes _input__fade-in {
-                from {
-                    opacity: 0;
-                }
-                100% {
-                    opacity: 1;
-                }
-            }
-
-            @keyframes fade {
-                from {
-                    opacity: 0.5;
-                }
-            }
-
-            ._input__fadeIn {
-                animation-name: :local(fade-in);
-            }
-
-            ._input__fadeIn {
-                animation: 2s :local(fade-in);
-            }
-
-            ._input__fadeIn {
-                animation: :local(fade-in) 2s;
-            }
-
-            :export {
-                fade-in: _input__fade-in;
-                fadeIn: _input__fadeIn;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_keywords_selector() {
-  test(
-    indoc! {r#"
-            :local(.constructor) {
-                color: green;
-            }
-
-            :local(.toString) {
-                color: red;
-            }
-        "#},
-    indoc! {r#"
-            ._input__constructor {
-                color: green;
-            }
-            __CSS_MODULE_LEXER_WHITESPACE__
-            ._input__toString {
-                color: red;
-            }
-            __CSS_MODULE_LEXER_WHITESPACE__
-            :export {
-                constructor: _input__constructor;
-                toString: _input__toString;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_multiple_classes() {
-  test(
-    indoc! {r#"
-            :local(.exportName) :local(.otherExport) {
-                color: green;
-            }
-
-            :local(.exportName):local(.otherExport) {
-                color: red;
-            }
-        "#},
-    indoc! {r#"
-            ._input__exportName ._input__otherExport {
-                color: green;
-            }
-
-            ._input__exportName._input__otherExport {
-                color: red;
-            }
-
-            :export {
-                exportName: _input__exportName;
-                otherExport: _input__otherExport;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_nested_class() {
-  test(
-    indoc! {r#"
-            :local(.exportName):not(:local(.otherExportName).global) {
-                color: green;
-            }
-
-            :local(.exportName):has(:local(.otherExportName), :local(.otherExportName2)) {
-                color: red;
-            }
-        "#},
-    indoc! {r#"
-            ._input__exportName:not(._input__otherExportName.global) {
-                color: green;
-            }
-
-            ._input__exportName:has(._input__otherExportName, ._input__otherExportName2) {
-                color: red;
-            }
-
-            :export {
-                exportName: _input__exportName;
-                otherExportName: _input__otherExportName;
-                otherExportName2: _input__otherExportName2;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_with_composes() {
-  test(
-    indoc! {r#"
-            :local(.otherClass) { background: red; } :local(.exportName) { compose-with: otherClass; color: green; }
-        "#},
-    indoc! {r#"
-            ._input__otherClass { background: red; } ._input__exportName {  color: green; }
-
-            :export {
-                otherClass: _input__otherClass;
-                exportName: _input__exportName _input__otherClass;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_with_composes_imported_class() {
-  // TODO: replace import value, should be `exportName: _lib_extender__exportName imported_otherClass;`
-  test(
-    indoc! {r#"
-            :import("./file.css") {
-                imported_otherClass: otherClass;
-            }
-            :local(.exportName) {
-                composes: imported_otherClass;
-                color: green;
-            }
-        "#},
-    indoc! {r#"
-            :import("./file.css") {
-                imported_otherClass: otherClass;
-            }
-            ._input__exportName {
-                __CSS_MODULE_LEXER_WHITESPACE__
-                color: green;
-            }
-
-            :export {
-                exportName: _input__exportName _input__imported_otherClass;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_with_global_composes() {
-  test(
-    indoc! {r#"
-            .otherClass { background: red; }
-            .andAgain { font-size: 2em; }
-            .aThirdClass { color: red; }
-            :local(.exportName) { compose-with: global(otherClass) global(andAgain); compose-with: global(aThirdClass); color: green; }
-        "#},
-    indoc! {r#"
-            .otherClass { background: red; }
-            .andAgain { font-size: 2em; }
-            .aThirdClass { color: red; }
-            ._input__exportName {   color: green; }
-
-            :export {
-                exportName: _input__exportName otherClass andAgain aThirdClass;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_with_multiple_composes() {
-  test(
-    indoc! {r#"
-            :local(.otherClass) { background: red; }
-            :local(.andAgain) { font-size: 2em; }
-            :local(.aThirdClass) { color: red; }
-            :local(.exportName) { compose-with: otherClass andAgain; compose-with: aThirdClass; color: green; }
-        "#},
-    indoc! {r#"
-            ._input__otherClass { background: red; }
-            ._input__andAgain { font-size: 2em; }
-            ._input__aThirdClass { color: red; }
-            ._input__exportName {   color: green; }
-
-            :export {
-                otherClass: _input__otherClass;
-                andAgain: _input__andAgain;
-                aThirdClass: _input__aThirdClass;
-                exportName: _input__exportName _input__otherClass _input__andAgain _input__aThirdClass;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn export_with_transitive_composes() {
-  test(
-    indoc! {r#"
-            :local(.aThirdClass) {
-                font-size: 2em;
-            }
-            :local(.otherClass) {
-                composes: aThirdClass;
-                background: red;
-            }
-            :local(.exportName) {
-                composes: otherClass;
-                color: green;
-            }
-        "#},
-    indoc! {r#"
-            ._input__aThirdClass {
-                font-size: 2em;
-            }
-            ._input__otherClass {
-                __CSS_MODULE_LEXER_WHITESPACE__
-                background: red;
-            }
-            ._input__exportName {
-                __CSS_MODULE_LEXER_WHITESPACE__
-                color: green;
-            }
-
-            :export {
-                aThirdClass: _input__aThirdClass;
-                otherClass: _input__otherClass _input__aThirdClass;
-                exportName: _input__exportName _input__otherClass _input__aThirdClass;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn ignore_custom_property_set() {
-  test(
-    indoc! {r#"
-            :root {
-                --title-align: center;
-                --sr-only: {
-                    position: absolute;
-                }
-            }
-        "#},
-    indoc! {r#"
-            :root {
-                --title-align: center;
-                --sr-only: {
-                    position: absolute;
-                }
-            }
-        "#},
-  );
-}
-
-#[test]
-fn multiple_composes() {
-  // TODO: replace import value
-  test(
-    indoc! {r#"
-            :import("path") {
-                i__i_a_0: a;
-                i__i_b_0: b;
-                i__i_c_0: c;
-                i__i_d_0: d;
-            }
-            :local(.class) {
-                composes: i__i_a_0 i__i_b_0, i__i_c_0, global(d) global(e), global(f), i__i_d_0;
-                color: red;
-            }
-        "#},
-    indoc! {r#"
-            :import("path") {
-                i__i_a_0: a;
-                i__i_b_0: b;
-                i__i_c_0: c;
-                i__i_d_0: d;
-            }
-            ._input__class {
-                __CSS_MODULE_LEXER_WHITESPACE__
-                color: red;
-            }
-
-            :export {
-                class: _input__class _input__i__i_a_0 _input__i__i_b_0 _input__i__i_c_0 d e f _input__i__i_d_0;
-            }
-        "#},
-  );
-}
-
-#[test]
-fn nested_rule() {
-  test(
-    indoc! {r#"
-            :root {
-                --test: {
-                    --test: foo;
-                    --bar: 1;
-                }
-            }
-        "#},
-    indoc! {r#"
-            :root {
-                --test: {
-                    --test: foo;
-                    --bar: 1;
-                }
-            }
-        "#},
-  );
-}
-
-#[test]
-fn nothing() {
-  test(
-    indoc! {r#"
-            .exportName {
-                color: green;
-            }
-
-            .exportName:hover {
-                color: red;
-            }
-
-            @media screen {
-                body {
-                    background: red;
-                }
-            }
-        "#},
-    indoc! {r#"
-            .exportName {
-                color: green;
-            }
-
-            .exportName:hover {
-                color: red;
-            }
-
-            @media screen {
-                body {
-                    background: red;
-                }
-            }
-        "#},
-  );
-}
-
-#[test]
-fn options_generate_scoped_name() {
-  test(
-    indoc! {r#"
-            :local(.exportName) {
-                color: green;
-            }
-
-            :local(.exportName):hover {
-                color: red;
-            }
-        "#},
-    indoc! {r#"
-            ._input__exportName {
-                color: green;
-            }
-
-            ._input__exportName:hover {
-                color: red;
-            }
-
-            :export {
-                exportName: _input__exportName;
             }
         "#},
   );
